[IMP] web: tree processor service
We create a new service tree_processor whose purpose is to ease the
processing of condition trees (~ domain). In that way, we are able to
simplify some code, notably treeFromDomain.
