+++ addons/web/static/src/search/search_model.js
@@ -6,14 +6,7 @@ import { DomainSelectorDialog } from "@web/core/domain_selector_dialog/domain_se
 import { _t } from "@web/core/l10n/translation";
 import { rpcBus } from "@web/core/network/rpc";
 import { evaluateExpr } from "@web/core/py_js/py";
-import { constructTreeFromDomain } from "@web/core/tree_editor/construct_tree_from_domain";
 import { domainFromTree } from "@web/core/tree_editor/domain_from_tree";
-import { treeFromDomain } from "@web/core/tree_editor/tree_from_domain";
-import {
-    useGetTreeDescription,
-    useGetTreeTooltip,
-    useMakeGetFieldDef,
-} from "@web/core/tree_editor/utils";
 import { user } from "@web/core/user";
 import { groupBy, sortBy } from "@web/core/utils/arrays";
 import { deepCopy } from "@web/core/utils/objects";
@@ -179,21 +172,16 @@ export class SearchModel extends EventBus {
         this.env = env;
         this.setup(services, args);
     }
-    /**
-     * @override
-     */
+
     setup(services) {
         // services
-        const { field: fieldService, name: nameService, orm, view, dialog } = services;
+        const { field: fieldService, orm, view, dialog, treeProcessor } = services;
         this.orm = orm;
         this.fieldService = fieldService;
         this.viewService = view;
+        this.treeProcessor = treeProcessor;
         this.dialog = dialog;
         this.orderByCount = false;
-
-        this.getDomainTreeDescription = useGetTreeDescription(fieldService, nameService);
-        this.getDomainTreeTooltip = useGetTreeTooltip(fieldService, nameService);
-        this.makeGetFieldDef = useMakeGetFieldDef(fieldService);
 
         // used to manage search items related to date/datetime fields
         this.referenceMoment = DateTime.local();
@@ -724,17 +712,17 @@ export class SearchModel extends EventBus {
             context = makeContext(contexts);
         }
 
-        const getFieldDef = await this.makeGetFieldDef(
+        const tree = await this.treeProcessor.treeFromDomain(
             this.resModel,
-            constructTreeFromDomain(domain)
+            domain,
+            !this.isDebugMode
         );
-        const tree = treeFromDomain(domain, { distributeNot: !this.isDebugMode, getFieldDef });
         const trees =
             !tree.negate && tree.value === "&" && tree.children.length > 0 ? tree.children : [tree];
         const promises = trees.map(async (tree) => {
             const [description, tooltip] = await Promise.all([
-                this.getDomainTreeDescription(this.resModel, tree),
-                this.getDomainTreeTooltip(this.resModel, tree, true),
+                this.treeProcessor.getDomainTreeDescription(this.resModel, tree),
+                this.treeProcessor.getDomainTreeTooltip(this.resModel, tree),
             ]);
             const preFilter = {
                 description,
