+++ addons/web/static/src/views/fields/domain/domain_field.js
@@ -7,10 +7,7 @@ import { _t } from "@web/core/l10n/translation";
 import { rpc } from "@web/core/network/rpc";
 import { EvaluationError } from "@web/core/py_js/py_builtin";
 import { registry } from "@web/core/registry";
-import { constructTreeFromDomain } from "@web/core/tree_editor/construct_tree_from_domain";
 import { domainContainsExpressions } from "@web/core/tree_editor/domain_contains_expressions";
-import { treeFromDomain } from "@web/core/tree_editor/tree_from_domain";
-import { useGetTreeDescription, useMakeGetFieldDef } from "@web/core/tree_editor/utils";
 import { useBus, useOwnedDialogs, useService } from "@web/core/utils/hooks";
 import { useRecordObserver } from "@web/model/relational_model/utils";
 import { SelectCreateDialog } from "@web/views/view_dialogs/select_create_dialog";
@@ -39,9 +36,8 @@ export class DomainField extends Component {
 
     setup() {
         this.orm = useService("orm");
-        this.getDomainTreeDescription = useGetTreeDescription();
-        this.makeGetFieldDef = useMakeGetFieldDef();
         this.notification = useService("notification");
+        this.treeProcessor = useService("tree_processor");
         this.getDefaultLeafDomain = useGetDefaultLeafDomain();
         this.addDialog = useOwnedDialogs();
 
@@ -166,13 +162,11 @@ export class DomainField extends Component {
         let promises = [];
         const domain = this.getDomain(props);
         try {
-            const getFieldDef = await this.makeGetFieldDef(
-                resModel,
-                constructTreeFromDomain(domain)
-            );
-            const tree = treeFromDomain(domain, { distributeNot: !this.env.debug, getFieldDef });
+            const tree = await this.treeProcessor.treeFromDomain(resModel, domain, !this.env.debug);
             const trees = !tree.negate && tree.value === "&" ? tree.children : [tree];
-            promises = trees.map((tree) => this.getDomainTreeDescription(resModel, tree));
+            promises = trees.map((tree) =>
+                this.treeProcessor.getDomainTreeDescription(resModel, tree)
+            );
         } catch (error) {
             if (error.data?.name === "builtins.KeyError" && error.data.message === resModel) {
                 // we don't want to support invalid models
